fix: PR #481レビュー対応完了 - structured_logger.py分割と技術的負債削減
主要変更:
- structured_logger.py: 1,313行→194行 (300行制限達成)
- バックアップファイル3件削除: 技術的負債40→37件
- Single Responsibility Principle適用
