+++ .github/pr_review_log.txt
@@ -1,0 +1,1 @@
+2025-07-16 08:13:34 - PR作成リマインダー表示
