+++ kumihan_formatter/core/utilities/log_optimization.py
@@ -1,0 +1,11 @@
+"""Log optimization module exports
+
+Single Responsibility Principle適用: ログ最適化モジュールの統合エクスポート
+Issue #476 Phase3対応 - log_optimization.py分割
+"""
+
+# Re-export classes from split modules
+from .log_performance import LogPerformanceOptimizer
+from .log_size_control import LogSizeController
+
+__all__ = ["LogPerformanceOptimizer", "LogSizeController"]
